initial commit
Formats the appearance of the puzzle given the selected puzzle letter. sets up the initial conditions to the puzzle. All letters in the crossword puzzle. Letters currently selected on the puzzle. Letters used in currently selected across and down clues. Across clue currently selected on the puzzle. Down clue currently selected on the puzzle. tells to run the init function once the page is fully loaded.